Split off common code into xmflib
