+++ x4lib.py
@@ -4,7 +4,6 @@ import os
 import os.path
 from copy import deepcopy
 from xml.etree import ElementTree
-from struct import calcsize, Struct
 import csv
 
 logger = logging.getLogger('x4.' + __name__)
@@ -97,33 +96,3 @@ class ModUtilMixin(object):
                 rows.append({h: d for h, d in zip(header, row[1:])})
 
 
-class StructException(Exception):
-    pass
-
-
-class StructObjBaseMeta(type):
-    def __init__(cls, name, bases, namespace):
-        super(StructObjBaseMeta, cls).__init__(name, bases, namespace)
-        cls.class_name = name
-        cls.fields = cls.fields.split(',')
-        cls.struct_len = calcsize(cls.struct_format)
-        cls.struct = Struct(cls.struct_format)
-
-
-class StructObjBase(object):
-    def __init__(self, **kwargs):
-        if set(self.fields) - set(kwargs):
-            raise StructException(f'Missing fields:\n\tpassed: {kwargs}, \n\trequired: {self.fields}')
-        self.__dict__ = kwargs
-
-    @classmethod
-    def from_stream(cls, stream, read_len=0):
-        obj = cls(**{k: v for k, v in zip(cls.fields, cls.struct.unpack(stream.read(cls.struct_len)))})
-        stream.seek(max(0, read_len - obj.struct_len), 1)
-        return obj
-
-    def to_stream(self):
-        return self.struct.pack(*(self.__dict__[f] for f in self.fields))
-
-    def __repr__(self):
-        return '%s(%s)' % (self.class_name, ', '.join('%s=%r' % (f, self.__dict__[f]) for f in self.fields))
